backend/requirements.txt
crewai>=0.140.0
fastapi>=0.100.0
uvicorn>=0.20.0
python-dotenv>=1.0.0
pydantic>=2.0.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-core>=0.2.0
httpx>=0.25.0
# Testing
pytest>=7.0.0
requests>=2.28.0

redis==5.0.1